230727 17:41 고친게 차아아암 많음 -명호

+++ Meevoo_Final/src/main/java/com/java/dto/SportDto.java
@@ -28,7 +28,7 @@ public class SportDto {
 	private String slope;
 	private int stotalstar;
 	private int stotalpeo;
-	private int savgstar;
+	private double savgstar;
 	private int stotalpick;
 	private String imgurl1;
 	private String imgurl2;
